Created MainMenu with buttons
Added a main menu containing a MarginContainer as a root( can be changed to a control node)
Using a Horizontal and vertical box container to split up the "UI" of the Menu and creating a layout.
I started working on the buttons so you can no see when a button is being hovered. There is no spacing currently between the buttons.
Added the quit functionality to the quit button so you can now close application on press