Added controllers and new main scene
Changed functionality so we base it all of 1 scene. THrough a levelController we load and unload each individual scene as a child of that node. So far you load mainmenu on startup, from mainmenu you can go to playerJoin/playMenu, There is no functionality yet in this menu except being able to return to mainmenu which works as of this commit