Added the rowing on multiple rows functionality.
Please  test and see how it feels. For me it seems kind of janky for now. it works. but not sure how good the coordination feels or if you can actually manage to go faster. Feels like it may be causing disruption with the bouncing on the "waves" we made.