Added variables for paddles
Added variables for paddles as to not clutter code with the same row.Row.GetChild(0) etc and reverted the change to what value to send to the GetPlayerInput function. This value is the controller that you are looking to get inputs from. By doing row.index you are forcing controller 0 to be on the front row  while having playerId makes it so the player who joins first is in the front row.