Created a game over menu
Made a game over menu as well as created some functionality to reset the state of a scene/ script through the levelmanager. Currently used to track if you have played and gone back to the main menu. This is used so we can empty the array which tracks the order players join and how many players have joined